training working

+++ model/encoder/encoderhierarchybase.py
@@ -73,7 +73,7 @@ class HierarchicalEncoder(nn.Module):
     
     def binary(self, x, bits):
         mask = 2**torch.arange(bits).to(x.device, x.dtype)
-        return x.bitwise_and(mask).ne(0).byte()
+        return x.bitwise_and(mask).ne(0).byte().to(dtype=torch.float)
     
     def binary_energy(self, x, lin_bits=20, sqrt_bits=20, log_bits=20):
         reps = int(np.floor(self.n_latent_nodes/(lin_bits+sqrt_bits+log_bits)))
